+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 1. [#2167](https://github.com/influxdata/chronograf/pull/2167): Add fractions of seconds to time field in csv export
 1. [#1077](https://github.com/influxdata/chronograf/pull/2087): Fix Chronograf requiring Telegraf's CPU and system plugins to ensure that all Apps appear on the HOST LIST page.
 1. [#2222](https://github.com/influxdata/chronograf/pull/2222): Fix template variables in dashboard query building.
+1. [#2291](https://github.com/influxdata/chronograf/pull/2291): Fix several kapacitor alert creation panics.
 
 ### Features
 ### UI Improvements